Nuevas rutas API, envío de correos y mejoras generales
- Se agregan endpoints para facturas por cliente, procesamiento masivo/individual de recordatorios y catálogos de clientes/ítems.
- Implementación de ApiResponse<T> para respuestas consistentes.
- Integración de envío de correos con SendGrid y modo consola para desarrollo, seleccionados automáticamente por configuración.
- Parámetros de SendGrid añadidos a appsettings.json y soporte para secretos de usuario.
- Refactor de casos de uso y repositorios para soportar nuevas operaciones.
- Ampliación de datos de prueba en mongo-init.js.
- Corrección de tipo en FinanceResume.Total.
- Eliminación de LoggerEmailService.
- Pruebas unitarias agregadas con Moq y xUnit.
- Habilitado CORS para facilitar integración con frontends.

+++ Domain/Entities/FinanceResume.cs
@@ -2,6 +2,6 @@
 {
     public class FinanceResume
     {
-        public decimal Total { get; set; }
+        public int Total { get; set; }
     }
 }
